+++ android/src/main/java/com/launchdarkly/RNLaunchDarklyModule.java
@@ -1,5 +1,5 @@
 
-package com.reactlibrary;
+package com.launchdarkly;
 
 import android.app.Application;
 import android.util.Log;
