+++ android/src/main/java/com/launchdarkly/RNLaunchDarklyPackage.java
@@ -1,5 +1,5 @@
 
-package com.reactlibrary;
+package com.launchdarkly;
 
 import java.util.Arrays;
 import java.util.Collections;
